Move couple class to dedicated file
